requirements.txt
django
django-redis
python-dotenv
daphne
channels
dnspython
requests
bs4
pillow
minio
